+++ packages/orq-rc/RELEASES.md
@@ -8099,3 +8099,13 @@ Based on:
 - [typescript v4.7.0-rc.53] packages/orq-rc
 ### Releases
 - [NPM v4.7.0-rc.53] https://www.npmjs.com/package/@orq-ai/node/v/4.7.0-rc.53 - packages/orq-rc
+
+## 2026-04-08 10:45:46
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.1 (2.879.6) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.7.0-rc.54] packages/orq-rc
+### Releases
+- [NPM v4.7.0-rc.54] https://www.npmjs.com/package/@orq-ai/node/v/4.7.0-rc.54 - packages/orq-rc